shared Ticket2Ticket function created
